Se añade las validaciones para el form de creación del libro

+++ Backend/src/models/book.js
@@ -24,7 +24,7 @@ const book = (sequelize) => {
             allowNull: false
         },
         price: {
-            type: DataTypes.INTEGER,
+            type: DataTypes.DECIMAL(7, 2),
             allowNull: false
         },
         image: {
